feat: Refactor settings screen and add SOS settings functionality
